Modified code to automatically store locs for each type of extension
Rather than forcing users to define what extensions they care about, we can just count LOCs for all types of files. This will also count lines for binary files that we don't care about, but it will also count lines of unknown formats.

Obviously this behaviour needs to be polished over time so that we can reach a better way of doing things, since counting lines of code only for specific formats is obviously also an useful thing to do.